refactor: moved handlers to file
